Backend/requirements.txt
aiofiles==23.2.1
alembic>=1.13.1
annotated-types>=0.7.0
anyio>=4.8.0
bcrypt==4.1.2
certifi==2025.1.31
cffi>=1.17.1
charset-normalizer==3.4.1
click>=8.1.8
colorama>=0.4.6
coverage==7.6.12
cryptography==41.0.7
dnspython==2.7.0
ecdsa>=0.19.0
email-validator==2.1.0.post1
environs==14.1.1
fastapi>=0.109.0
fastapi_cors==0.0.6
greenlet>=3.1.0
h11>=0.14.0
httpcore==1.0.7
httpx==0.26.0
idna>=3.6
iniconfig==2.0.0
Mako>=1.3.9
MarkupSafe>=3.0.2
marshmallow==3.26.1
mysql-connector-python==8.0.33
packaging==24.2
passlib>=1.7.4
pluggy==1.5.0
prometheus_client==0.21.1
protobuf==3.20.3
pyasn1==0.6.1
pycparser>=2.22
pydantic>=2.0.0
pydantic-settings==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
setuptools>=69.0.3
six==1.16.0
sniffio>=1.3.1
SQLAlchemy>=1.4.0
starlette>=0.35.1
typing_extensions>=4.9.0
urllib3==2.3.0
uvicorn>=0.27.0
wheel>=0.42.0
